Long userId cannot be in List<Integer> (and throws an error because it can't be converted)

+++ backend/src/main/java/com/analysetool/repositories/WPUserMetaRepository.java
@@ -72,10 +72,10 @@ public interface WPUserMetaRepository extends JpaRepository<WPUserMeta, Long> {
     List<Long> getUserIdsByTag(String tag);
 
     @Query("SELECT DISTINCT(u.userId) FROM WPUserMeta u WHERE u.key = 'profile_tags'")
-    List<Integer> getAllUserIdsWithTags();
+    List<Long> getAllUserIdsWithTags();
 
     @Query("SELECT u.value FROM WPUserMeta u WHERE u.key = 'profile_tags' AND u.userId IN :list")
-    List<String> getAllUserTagRowsInList(List<Integer> list);
+    List<String> getAllUserTagRowsInList(List<Long> list);
 
     @Query("SELECT count(DISTINCT p.userId) FROM WPUserMeta p WHERE p.key='profile_tags' AND p.value <> 'a:0:{}'")
     Integer getTotalCountOfUsersWithTag();
